+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -325,6 +325,8 @@ object HybrisConstants {
     const val SCHEMA_COCKPIT_NG_WIDGETS = "http://www.hybris.com/schema/cockpitng/widgets.xsd"
     const val SCHEMA_COCKPIT_NG_CONFIG = "http://www.hybris.com/cockpit/config"
 
+    const val WS_RELATED = "wsRelated"
+
     val DEFAULT_JUNK_FILE_NAMES = listOf(
         ".classpath",
         ".directory",