+++ CHANGELOG.md
@@ -72,7 +72,7 @@
 - Include `backoffice.web` libraries in the `Backoffice Library` [#734](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/734)
 
 ### `beans.xml` inspection rules
-- new inspection that suggests to add the hint "wsRelated" for generating Swagger documentation annotation [#752](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/752)
+- New inspection that suggests to add the hint "wsRelated" for generating Swagger documentation annotation [#752](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/752)
 
 ### Fixes
 - `SAP Commerce` tool window sometimes appears without any content [#725](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/725)